Update Workspaces_in_SAP_Business_Application_Studio_0919ce1.md
Line number 23: I am not sure about this change. Not really sure whether launch is a separate word (or) it should be launch settings.
If launch is a separate word, I cannot understand what it means. Is "launch" a separate resource of a project?
Please ignore if my suggestion is wrong.